Chat-bot/requirements.txt
streamlit
pandas
faiss-cpu
sentence-transformers
transformers
python-dotenv
openai==0.28
